Update DomainSync.php
With the latest versions of the module the nextDueDateOffset cannot be 0. This is parsed as an empty integer and thus defaulted to 14 days. This is not according to the documentation when in case the nextDueDateOffset is not set, but updateNextDueDate is, it will update the value equal to the expiry date.

https://github.com/openprovider/OP-WHMCS7/blob/master/docs/configure_openprovider_cron_sync.md